LPS-161451 Show tooltip with the shortcut
