+++ modules/apps/portal-language/portal-language-lang/src/main/resources/content/Language.properties
@@ -17170,6 +17170,7 @@ toggle-controls=Toggle Controls
 toggle-default=Toggle Default
 toggle-dropdown=Toggle Dropdown
 toggle-navigation=Toggle Navigation
+toggle-preview=Toggle Preview
 token=Token
 token-configuration-name=Token Based SSO
 token-connection-timeout=Token Connection Timeout in Milliseconds
